Remove gw-generator dependency from gw-store
