+++ crates/common/src/traits.rs
@@ -1,0 +1,9 @@
+use crate::H256;
+use gw_types::{bytes::Bytes, packed::Script};
+
+pub trait CodeStore {
+    fn insert_script(&mut self, script_hash: H256, script: Script);
+    fn get_script(&self, script_hash: &H256) -> Option<Script>;
+    fn insert_data(&mut self, data_hash: H256, code: Bytes);
+    fn get_data(&self, data_hash: &H256) -> Option<Bytes>;
+}
