+++ crates/generator/Cargo.toml
@@ -9,6 +9,8 @@ edition = "2018"
 [dependencies]
 gw-types = { path = "../types" }
 gw-common = { path = "../common" }
+gw-config = { path = "../config" }
+gw-store = { path = "../store" }
 anyhow = "1.0"
 blake2b-rs = "0.2"
 ckb-vm = { git = "https://github.com/nervosnetwork/ckb-vm.git", tag = "0.19.1", features = ["detect-asm"] }
