+++ crates/generator/src/dummy_state.rs
@@ -1,8 +1,8 @@
-use crate::traits::CodeStore;
 use gw_common::{
     error::Error,
     smt::{default_store::DefaultStore, H256, SMT},
     state::State,
+    CodeStore,
 };
 use gw_types::{bytes::Bytes, packed::Script};
 use std::collections::HashMap;
