+++ crates/generator/src/lib.rs
@@ -3,11 +3,10 @@
 
 pub mod account_lock_manage;
 pub mod backend_manage;
-pub mod builtin_scripts;
 pub mod dummy_state;
 pub mod error;
 pub mod generator;
-pub mod sudt;
+pub mod genesis;
 pub mod syscalls;
 #[cfg(test)]
 mod tests;
@@ -18,11 +17,3 @@ mod types;
 pub use error::Error;
 pub use generator::Generator;
 pub use types::*;
-
-pub(crate) fn code_hash(data: &[u8]) -> gw_common::H256 {
-    let mut hasher = gw_common::blake2b::new_blake2b();
-    hasher.update(data);
-    let mut code_hash = [0u8; 32];
-    hasher.finalize(&mut code_hash);
-    code_hash.into()
-}
