+++ crates/generator/src/syscalls/mod.rs
@@ -1,4 +1,3 @@
-use crate::traits::CodeStore;
 use crate::types::RunResult;
 use ckb_vm::{
     memory::Memory,
@@ -12,7 +11,7 @@ use gw_common::{
         build_account_field_key, build_script_hash_to_account_id_key, State, GW_ACCOUNT_NONCE,
         GW_ACCOUNT_SCRIPT_HASH,
     },
-    H256,
+    CodeStore, H256,
 };
 use gw_types::{
     bytes::Bytes,
