+++ crates/generator/src/tests/scripts/sudt.rs
@@ -1,16 +1,13 @@
 use super::new_block_info;
 use crate::{account_lock_manage::AccountLockManage, Generator};
-use crate::{
-    backend_manage::BackendManage, builtin_scripts::SUDT_VALIDATOR_CODE_HASH,
-    dummy_state::DummyState,
-};
+use crate::{backend_manage::BackendManage, dummy_state::DummyState};
 use crate::{
     error::TransactionError,
     traits::{CodeStore, StateExt},
 };
 use core::panic;
 use gw_common::state::State;
-use gw_common::{h256_ext::H256Ext, H256};
+use gw_common::{builtin_scripts::SUDT_VALIDATOR_CODE_HASH, h256_ext::H256Ext, H256};
 use gw_types::{
     packed::{BlockInfo, RawL2Transaction, SUDTArgs, SUDTQuery, SUDTTransfer, Script},
     prelude::*,
