+++ crates/store/Cargo.toml
@@ -11,7 +11,6 @@ gw-db = { path = "../db" }
 gw-types = { path = "../types" }
 gw-config = { path = "../config" }
 gw-common = { path = "../common" }
-gw-generator = { path = "../generator" }
 anyhow = "1.0"
 thiserror = "1.0"
 parking_lot = "0.11"
